Extract a DestinationSelectorComponent
From https://github.com/sul-dlss/sul-requests/pull/1638\#discussion_r1293674040
