+++ app/views/requests/_destination_and_needed_date.html.erb
@@ -1,5 +1,5 @@
 <div class="alert alert-warning destination-note-callout">
-  <%= render partial: 'destination_selector', locals: { f: f } %>
+  <%= render DestinationSelectorComponent.new(form: f) %>
 
   <%= render partial: 'needed_date', locals: { f: f } %>
 </div>
