--- app/views/requests/_destination_selector.html.erb
@@ -1,4 +1,0 @@
-<div data-scheduler-lookup-url='<%= paging_schedule_path(origin: current_request.paging_origin_library) %>'>
-  <%= select_for_pickup_destinations(f).html_safe %>
-  <%= render partial: 'scheduler_text', locals: { f: f } %>
-</div>
