+++ spec/views/shared/_request_status_information.html.erb_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe 'shared/_request_status_information.html.erb' do
     context 'when there is no delivery destination (e.g. for a scan)' do
       let(:request) { build_stubbed(:scan, :without_validations, :with_item_title, user:) }
 
-      it "doesn't display the 'Deliver to' field" do
+      it "doesn't display the 'Will be delivered to' field" do
         render
         expect(rendered).not_to have_content('Will be delivered to')
       end
@@ -22,7 +22,7 @@ RSpec.describe 'shared/_request_status_information.html.erb' do
     context 'when there is a delivery destination' do
       let(:request) { build_stubbed(:page_mp_mediated_page, user:, barcodes: ['12345678']) }
 
-      it "displays the 'Deliver to' field" do
+      it "displays the 'Will be delivered to' field" do
         render
         expect(rendered).to have_content('Will be delivered to')
       end
